test pid lidar : 3

+++ catkin_ws/src/pid_lidar/include/pid/controller_lidar.h
@@ -97,7 +97,7 @@ std::string topic_from_plant;
 std::string setpoint_topic;
 std::string node_name = "pid_node";
 
-std_msgs::UInt16 control_msg;
+std_msgs::Float64 control_msg;
 std_msgs::Float64 state_msg;
 
 // Diagnostic objects
